backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.3
openai==1.65.1
pandas==2.2.3
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
fastapi
uvicorn
sqlalchemy
pydantic
